% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmerge.R
\name{diffmerge}
\alias{diffmerge}
\title{A function}
\usage{
diffmerge(filelist, path)
}
\arguments{
\item{file}{A string}

\item{expr}{A variable in the dataframe}
}
\value{
A dataframe
}
\description{
A function
}
