% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMED.R
\name{findMED}
\alias{findMED}
\title{A function}
\usage{
findMED(cohort, data, medname)
}
\arguments{
\item{file}{A string}

\item{expr}{A variable in the dataframe}
}
\value{
A dataframe
}
\description{
A function
}
