% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readacleansheet.R
\name{readacleansheet}
\alias{readacleansheet}
\title{A function}
\usage{
readacleansheet(filename, sheetname)
}
\arguments{
\item{file}{A string}

\item{expr}{A variable in the dataframe}
}
\value{
A dataframe
}
\description{
A function
}
