% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readaexcel.R
\name{readaexcel}
\alias{readaexcel}
\title{A function}
\usage{
readaexcel(file)
}
\arguments{
\item{file}{A string}

\item{expr}{A variable in the dataframe}
}
\value{
A dataframe
}
\description{
A function
}
