% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnmerge.R
\name{readnmerge}
\alias{readnmerge}
\title{A function}
\usage{
readnmerge(html = F)
}
\arguments{
\item{file}{A string}

\item{expr}{A variable in the dataframe}
}
\value{
A dataframe
}
\description{
A function
}
