% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shinyreadonemerge.R
\name{Shinyreadonemerge}
\alias{Shinyreadonemerge}
\title{A function}
\usage{
Shinyreadonemerge(path, keyword, xls = F, writexlsx = F, writesav = F)
}
\arguments{
\item{file}{A string}

\item{expr}{A variable in the dataframe}
}
\value{
A dataframe
}
\description{
A function
}
